Added the plugin.yml
The plugin.yml contains all the necessary information about the plugin.
It contains the name of the plugin, the location of the class
representing the plugin (DowerTefence) and finally the version of the
plugin.

The Plugin can now successfully be run by the server, even though it
does nothing. First build the project (clean install), then copy the
created jar file (found in the target folder) into the plugin folder of
the server. When the server gets launched it should specify somewhere in
the console that the plugin has been loaded successfully.